requirements.txt
pandas
awscli
pyarrow
s3fs
fsspec
pytorch-lightning
transformers
tabulate
tensorflow>=2.5.0
Pillow
gensim>=4.0.1
matplotlib